Add files via upload
I learned about the system unit and its components through the jigsaw activity. I discovered that there are six major components essential for a computer to function. If we had more time, it would be interesting and fun to create a 3D model of the system unit to present to the class. This could help visualize the internal parts more clearly and make the learning experience more engaging.